fix: Correct data types in weather models and add unit tests
Corrects data model inconsistencies and adds unit tests for the `WeatherApiService` to verify data fetching and deserialization logic.

This commit includes the following changes:

-   **Data Model Fixes**:
    -   Updates `DailyForecastWeatherData` to change `lat` and `lon` from `Double?` to `String?`.
    -   Modifies `DailyData` to change `dewPoint` from `Int?` to `Double?`.

-   **Unit Testing**:
    -   Adds a new test file, `WeatherApiServiceTest.kt`, to the `networking` module.
    -   Includes tests to verify successful fetching and deserialization of `DailyForecastWeatherData` and to confirm that deserialization fails with an incorrect `Content-Type`.
    -   Integrates `Mokkery`, `kotlinx-coroutines-test`, and `ktor-client-mock` for testing purposes.

-   **Mocking Support**:
    -   Applies the `OpenClassForMocking` annotation to `HttpClientFactory` to enable mocking in tests.

+++ feature-weather/networking/src/commonMain/kotlin/com/charly/networking/model/DailyData.kt
@@ -29,7 +29,7 @@ data class DailyData(
     @SerialName("humidity")
     val humidity: Int? = null,
     @SerialName("dew_point")
-    val dewPoint: Int? = null,
+    val dewPoint: Double? = null,
     @SerialName("wind_speed")
     val windSpeed: Double? = null,
     @SerialName("wind_deg")
